+++ analysis/max/max.h
@@ -3,4 +3,4 @@
 #include "../computation_struct.h"
 #include "max_settings.h"
 
-__device__ void MAX(Computation* data, int variation, void(*finiteDifferenceScheme)(numb*, numb*, numb*));
+__device__ void MAX(Computation* data, uint64_t variation, void(*finiteDifferenceScheme)(numb*, numb*, numb*));
